+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/core/FlysqlCondition.java
@@ -248,4 +248,12 @@ public interface FlysqlCondition<T> {
      * @since 0.0.1
      */
     T isNotDeleted();
+
+    /**
+     * isDeleted()
+     *
+     * @return se
+     * @since 0.0.1
+     */
+    T isDeleted();
 }